Move logic to select the toolchain for a document from `SourceKitLSPServer` to `BuildSystemManager`
This allows us to determine the toolchain to use during background indexing. It also moves toolchain selection closer to the build system, which is good because when we support multiple toolchains for a single workspace, the build system is what decides which toolchain to use for which document.
